requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-extensions==4.1
idna==3.10
ijson==3.4.0
numpy==2.3.2
polars==1.32.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
